+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [develop](https://github.com/sprsquish/blather/compare/master...develop)
+  * Bugfix(benlangfeld): Ensure that presence nodes which are both status and subscription may be responded to
 
 # [v0.7.1](https://github.com/sprsquish/blather/compare/v0.7.0...v0.7.1) - [2012-04-29](https://rubygems.org/gems/blather/versions/0.7.1)
   * Documentation updates